# docs/openapi.yaml
openapi: 3.1.0
info:
  title: UPP Gateway API
  version: 0.1.0
  description: |
    Universal Prediction Protocol Gateway API provides unified access to prediction markets across multiple providers via a single API.

    ## Authentication

    The API uses two authentication methods:
    - **ApiKeyAuth**: X-API-Key header for API key-based authentication
    - **BearerAuth**: Bearer token in Authorization header for JWT-based authentication

    Protected endpoints require authentication. Public endpoints are rate-limited but do not require credentials.

    ## Rate Limiting

    All endpoints are rate-limited. Rate limit information is returned in response headers:
    - `X-RateLimit-Limit`: Maximum requests per minute
    - `X-RateLimit-Remaining`: Remaining requests in current window
    - `Retry-After`: Seconds to wait before retrying (on 429 responses)

  contact:
    name: UPP Team
    url: https://github.com/upp-labs
    email: support@upp.dev
  license:
    name: Apache-2.0
    url: https://opensource.org/licenses/Apache-2.0
servers:
  - url: http://localhost:8080
    description: Local development server
  - url: http://gateway:8080
    description: Docker containerized gateway
  - url: https://api.upp.dev
    description: Production server
...
paths:
...
  /.well-known/upp:
    get:
      operationId: getWellKnown
      summary: UPP well-known configuration
      description: Returns well-known configuration for UPP discovery
      tags: [Discovery]
      responses:
        '200':
          description: Well-known configuration
          content:
            application/json:
              schema:
                type: object
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /upp/v1/discovery/manifest/{provider}:
    get:
      operationId: getDiscoveryManifest
      summary: Get provider manifest
      description: Get the manifest for a specific prediction market provider
      tags: [Discovery]
      parameters:
        - name: provider
          in: path
          required: true
          description: Provider name (e.g., 'kalshi.com', 'polymarket.com')
          schema:
            type: string
      responses:
        '200':
          description: Provider manifest
          content:
            application/json:
              schema:
                type: object
        '400':
          description: Invalid provider name
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Provider not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /upp/v1/discovery/negotiate:
    post:
      operationId: discoverNegotiate
      summary: Provider negotiation
      description: Negotiate connection parameters with a provider
      tags: [Discovery]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                provider:
                  type: string
                  description: Provider name
                parameters:
                  type: object
                  description: Negotiation parameters
      responses:
        '200':
          description: Negotiation successful
          content:
            application/json:
              schema:
                type: object
        '400':
          description: Invalid negotiation parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /upp/v1/markets:
    get:
      operationId: listMarkets
      summary: List markets
      description: List all available prediction markets
      tags: [Markets]
      parameters:
        - name: provider
          in: query
          description: Filter by provider name
          schema:
            type: string
        - name: status
          in: query
          description: Filter by market status (open, closed, etc.)
          schema:
            type: string
        - name: category
          in: query
          description: Filter by market category
          schema:
            type: string
        - name: limit
          in: query
          description: Maximum number of results (default 50)
          schema:
            type: integer
            minimum: 1
            maximum: 1000
        - name: cursor
          in: query
          description: Pagination cursor
          schema:
            type: string
      responses:
        '200':
          description: List of markets
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MarketsResponse'
        '400':
          description: Invalid parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Rate limited
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /upp/v1/markets/search:
    get:
      operationId: searchMarkets
      summary: Search markets
      description: Search for markets by title or description
      tags: [Markets]
      parameters:
        - name: q
          in: query
          required: false
          description: Search query
          schema:
            type: string
        - name: provider
          in: query
          description: Filter by provider
          schema:
            type: string
        - name: category
          in: query
          description: Filter by category
          schema:
            type: string
        - name: limit
          in: query
          description: Maximum number of results
          schema:
            type: integer
            minimum: 1
            maximum: 1000
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchResponse'
        '400':
          description: Invalid search parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /upp/v1/markets/{marketId}:
    get:
      operationId: getMarket
      summary: Get market details
      description: Get detailed information about a specific market
      tags: [Markets]
      parameters:
        - name: marketId
          in: path
          required: true
          description: Market identifier
          schema:
            type: string
      responses:
        '200':
          description: Market details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MarketResponse'
        '404':
          description: Market not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /upp/v1/markets/{marketId}/orderbook:
    get:
      operationId: getOrderbook
      summary: Get market orderbook
      description: Get the current orderbook for a market
      tags: [Orderbook]
      parameters:
        - name: marketId
          in: path
          required: true
          description: Market identifier
          schema:
            type: string
      responses:
        '200':
          description: Market orderbook
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderbookResponse'
        '404':
          description: Market not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /upp/v1/markets/{marketId}/orderbook/merged:
    get:
      operationId: getMergedOrderbook
      summary: Get merged orderbook
      description: Get merged orderbook across all providers for a market
      tags: [Orderbook]
      parameters:
        - name: marketId
          in: path
          required: true
          description: Market identifier
          schema:
            type: string
      responses:
        '200':
          description: Merged orderbook
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderbookResponse'
        '404':
          description: Market not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /upp/v1/arbitrage:
    get:
      operationId: listArbitrageOpportunities
      summary: List arbitrage opportunities
      description: Get list of current arbitrage opportunities across markets and providers
      tags: [Arbitrage]
      parameters:
        - name: min_profit
          in: query
          description: Filter opportunities by minimum profit percentage
          schema:
            type: number
            format: float
      responses:
        '200':
          description: Arbitrage opportunities
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArbitrageListResponse'
        '429':
          description: Rate limited
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /upp/v1/arbitrage/history:
    get:
      operationId: getArbitrageHistory
      summary: Arbitrage history
      description: Get historical arbitrage trades and opportunities
      tags: [Arbitrage]
      parameters:
        - name: limit
          in: query
          description: Maximum number of historical entries
          schema:
            type: integer
            minimum: 1
            maximum: 1000
      responses:
        '200':
          description: Arbitrage history
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArbitrageHistoryResponse'

  # ============================================================================
  # PRICE HISTORY & CANDLES
  # ============================================================================
  /upp/v1/markets/{marketId}/candles:
    get:
      operationId: getCandles
      summary: Get candlestick data
      description: Get OHLCV candlestick data for a market outcome
      tags: [Candles]
      parameters:
        - name: marketId
          in: path
          required: true
          description: Market identifier
          schema:
            type: string
        - name: outcome_id
          in: query
          description: Outcome identifier (if not specified, returns all outcomes)
          schema:
            type: string
        - name: resolution
          in: query
          description: Candle resolution (1m, 5m, 15m, 1h, 4h, 1d)
          schema:
            type: string
            enum: [1m, 5m, 15m, 1h, 4h, 1d]
        - name: from
          in: query
          description: Start timestamp (ISO 8601)
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          description: End timestamp (ISO 8601)
          schema:
            type: string
            format: date-time
        - name: limit
          in: query
          description: Maximum number of candles
          schema:
            type: integer
            minimum: 1
            maximum: 5000
      responses:
        '200':
          description: Candlestick data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CandlesResponse'
        '400':
          description: Invalid parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Market not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /upp/v1/markets/{marketId}/candles/latest:
    get:
      operationId: getLatestCandle
      summary: Get latest candle
      description: Get the latest candlestick for a market outcome
      tags: [Candles]
      parameters:
        - name: marketId
          in: path
          required: true
          description: Market identifier
          schema:
            type: string
        - name: outcome_id
          in: query
          description: Outcome identifier
          schema:
            type: string
        - name: resolution
          in: query
          description: Candle resolution
          schema:
            type: string
            enum: [1m, 5m, 15m, 1h, 4h, 1d]
      responses:
        '200':
          description: Latest candle
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LatestCandleResponse'
        '404':
          description: Market not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /upp/v1/price-index/stats:
    get:
      operationId: getPriceIndexStats
      summary: Get price index statistics
      description: Get aggregated price index statistics across all markets
      tags: [Price Index]
      parameters:
        - name: index_id
          in: query
          description: Specific index identifier
          schema:
            type: string
      responses:
        '200':
          description: Price index statistics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PriceIndexStatsResponse'

  # ============================================================================
  # RESOLUTION & SETTLEMENT
  # ============================================================================
  /upp/v1/resolutions/{marketId}:
    get:
      operationId: getResolution
      summary: Get market resolution
      description: Get resolution data for a specific market
      tags: [Resolution]
      parameters:
        - name: marketId
          in: path
          required: true
          description: Market identifier
          schema:
            type: string
      responses:
        '200':
          description: Resolution data
          content:
            application/json:
              schema:
                type: object
        '404':
          description: Market not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /upp/v1/resolutions:
    get:
      operationId: listResolutions
      summary: List resolutions
      description: List all available resolutions
      tags: [Resolution]
      parameters:
        - name: status
          in: query
          description: Filter by resolution status
          schema:
            type: string
      responses:
        '200':
          description: List of resolutions
          content:
            application/json:
              schema:
                type: object
                properties:
                  resolutions:
                    type: array
                    items:
                      type: object
...
  /upp/v1/backtest/run:
    post:
      operationId: runBacktest
      summary: Run backtest
      description: Execute a backtest of a strategy against historical market data
      tags: [Backtest]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RunBacktestRequest'
      responses:
        '200':
          description: Backtest completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BacktestResponse'
        '400':
          description: Invalid backtest parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /upp/v1/backtest/compare:
    post:
      operationId: compareStrategies
      summary: Compare strategies
      description: Compare multiple strategies against the same market
      tags: [Backtest]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompareStrategiesRequest'
      responses:
        '200':
          description: Comparison completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompareStrategiesResponse'
        '400':
          description: Invalid comparison parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ============================================================================
  # FEEDS
  # ============================================================================
...
  /upp/v1/ws:
    get:
      operationId: websocketUpgrade
      summary: WebSocket upgrade
      description: Upgrade to WebSocket connection for real-time data streaming
      tags: [WebSocket]
      security:
        - ApiKeyAuth: []
        - BearerAuth: []
      responses:
        '101':
          description: WebSocket connection established
        '400':
          description: Invalid WebSocket upgrade request
        '401':
          description: Unauthorized
...
components:
...
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API Key authentication using X-API-Key header

    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Bearer token authentication using JWT